+++ vm/src/types/instance_definitions/segment_arena_instance_def.rs
@@ -1,0 +1,4 @@
+use serde::Serialize;
+
+#[derive(Serialize, Clone, Debug, PartialEq, Default)]
+pub(crate) struct SegmentArenaInstanceDef;